requirements.txt
streamlit>=1.32
anthropic>=0.39
openpyxl>=3.1
python-docx>=1.1
pandas>=2.0
pdfplumber>=0.10
Pillow>=10.0
google-api-python-client>=2.120
google-auth>=2.28
